chore: segregate command extractor code
